review(sync): SaturatingSub, ulong skipLastN/maxHeaders/fastSyncLag, strip noise
PowForwardHeaderProvider: SaturatingSub for chunk arithmetic (headNum - 1,
HeadNumber - skipLastN, currentNumber - ancestorJump, bestSuggestedNumber -
MaxReorganizationLength, currentNumber - 1028). maxHeaders/skipLastN/
headersToRequest now ulong (cascades through IForwardHeaderProvider,
IForwardSyncController, BlockDownloader.PrepareRequest, FastSyncFeed).

PosForwardHeaderProvider: ulong signature at the entry then narrows once to
int internally so the array/buffer arithmetic stays int-typed.

SyncServer: SaturatingSub on the MaxReorgLength gate and the seal-validation
hint window; drop the dead `number >= 0` guard (number is ulong); drop the
"in the range of [Head - MaxReorganizationLength, Head]" stale comment.

PeerInfo.ShouldNotifyNewRange: takes ulong instead of long; the two private
last-notified fields cascade. SyncServer call site drops the (long) casts.

ParallelSync: SyncProgressResolver FindBestProcessedBlock collapses
`is { } n ? n : ulong.MaxValue` into `?? ulong.MaxValue`; SyncPivot getter
collapses the trivial deconstruction. MultiSyncModeSelector: drop the long
"Safe: …" comment that re-stated the ulong wrap behaviour.

FullStateFinder: strip the "Block numbers are ulong…" and "startHeader.Number
is ulong" comments; flatten the nested-if guard against underflow.

SyncReport: SaturatingSub on the numHeadersToDownload subtraction.

BetterPeerStrategyExtensions, ITotalDifficultyStrategy, TreeSync,
SyncPeerPool: drop the leftover "cast is safe / Number is ulong" comments.

Tests: BlockDownloaderTests / ForwardHeaderProviderTests /
PosForwardHeaderProviderCacheTests / BlockDownloaderTests.BlockAccessLists
follow the new ulong signatures.

+++ src/Nethermind/Nethermind.Merge.Plugin.Test/Synchronization/PosForwardHeaderProviderCacheTests.cs
@@ -66,7 +66,7 @@ public class PosForwardHeaderProviderCacheTests
     public void TearDown() => _provider.UnsubscribeForTest();
 
     private Task<IOwnedReadOnlyList<BlockHeader?>?> Get(int skip = 0, int max = Requested) =>
-        _provider.GetBlockHeaders(skip, max, CancellationToken.None);
+        _provider.GetBlockHeaders((ulong)skip, (ulong)max, CancellationToken.None);
 
     private void RaiseMainChainUpdate(params Block[] blocks)
     {
