+++ src/Nethermind/Nethermind.Synchronization/Blocks/FastSyncFeed.cs
@@ -16,7 +16,7 @@ namespace Nethermind.Synchronization.Blocks
 
         private DownloaderOptions BuildOptions() => DownloaderOptions.Insert | DownloaderOptions.WithReceipts;
 
-        public override Task<BlocksRequest> PrepareRequest(CancellationToken token = default) => forwardSyncController.PrepareRequest(BuildOptions(), (int)syncConfig.StateMinDistanceFromHead, token);
+        public override Task<BlocksRequest> PrepareRequest(CancellationToken token = default) => forwardSyncController.PrepareRequest(BuildOptions(), syncConfig.StateMinDistanceFromHead, token);
 
         public override SyncResponseHandlingResult HandleResponse(BlocksRequest response, PeerInfo peer = null) => forwardSyncController.HandleResponse(response, peer);
